Fix: firebase-auth
