+++ .config/firebase/firestore.rules
@@ -33,6 +33,7 @@ service cloud.firestore {
       // 書き込みは管理者のみ許可
       allow write: if request.auth != null;
     }
+    
     // その他のドキュメントへのアクセスを拒否
     match /{document=**} {
       allow read, write: if false;
