+++ .config/firebase/storage.rules
@@ -1,8 +1,29 @@
 rules_version = '2';
 service firebase.storage {
   match /b/{bucket}/o {
+    // Profiles画像用のルール
+    match /profiles/{allImages=**} {
+      // 読み取りは全ユーザーに許可
+      allow read: if true;
+      // 書き込みは認証済みユーザーのみ許可
+      allow write: if request.auth != null 
+        && request.resource.size < 5 * 1024 * 1024  // 5MB以下
+        && request.resource.contentType.matches('image/.*');  // 画像のみ
+    }
+
+    // Works画像用のルール
+    match /works/{allImages=**} {
+      // 読み取りは全ユーザーに許可
+      allow read: if true;
+      // 書き込みは認証済みユーザーのみ許可
+      allow write: if request.auth != null
+        && request.resource.size < 5 * 1024 * 1024  // 5MB以下
+        && request.resource.contentType.matches('image/.*');  // 画像のみ
+    }
+
+    // その他のファイルへのアクセスを拒否
     match /{allPaths=**} {
-      allow read, write
+      allow read, write: if false;
     }
   }
 }
