+++ package.json
@@ -15,7 +15,6 @@
 		"firebase": "^11.8.1",
 		"firebase-admin": "^13.4.0",
 		"next": "15.3.1",
-		"next-auth": "^5.0.0-beta.28",
 		"react": "^19.0.0",
 		"react-dom": "^19.0.0",
 		"react-hot-toast": "^2.5.2",
