+++ src/app/admin/page.tsx
@@ -1,9 +1,11 @@
+'use client';
+
 import { AdminLayout } from "@/components/layout/AdminLayout";
 import Link from "next/link";
-import { redirect } from "next/navigation";
-import { auth } from "@/auth";
+import { useRouter } from "next/navigation";
 import { FaUser, FaClipboard, FaToolbox } from "react-icons/fa";
 import type { IconType } from "react-icons";
+import { useAuth } from "@/components/auth/AuthProvider";
 
 interface AdminMenu {
 	id: string;
@@ -37,12 +39,24 @@ const ADMIN_MENUS: AdminMenu[] = [
 	},
 ];
 
-export default async function AdminPage() {
+export default function AdminPage() {
 	// サーバーサイドで認証チェック
-	const session = await auth();
-
-	if (!session) {
-		redirect("/admin/login");
+	const router = useRouter();
+	const { user, loading } = useAuth();	
+	// 認証状態をチェック
+	if (loading) {
+	  return (
+	    <AdminLayout>
+	      <div className="flex min-h-screen items-center justify-center">
+	        <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-gray-900" />
+	      </div>
+	    </AdminLayout>
+	  );
+	}	
+	// 未認証の場合はログインページへリダイレクト
+	if (!user) {
+	  router.replace("/admin/login");
+	  return null;
 	}
 
 	return (
