+++ src/components/admin/AuthGuard.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useSession } from "next-auth/react";
+import { useAuth } from "@/components/auth/AuthProvider";
 import { redirect } from "next/navigation";
 
 interface AuthGuardProps {
@@ -8,19 +8,18 @@ interface AuthGuardProps {
 }
 
 export function AuthGuard({ children }: AuthGuardProps) {
-	const { status } = useSession();
-
-	if (status === "loading") {
-		return (
-			<div className="flex min-h-screen items-center justify-center">
-				<div className="h-12 w-12 animate-spin rounded-full border-b-2 border-blue-600"></div>
-			</div>
-		);
+	const { user, loading } = useAuth();
+	if (loading) {
+	  return (
+	    <div className="flex min-h-screen items-center justify-center">
+	      <div className="h-12 w-12 animate-spin rounded-full border-b-2 border-blue-600"></div>
+	    </div>
+	  );
 	}
 
-	if (status === "unauthenticated") {
-		redirect("/admin/login");
+	if (!user) {
+	  redirect("/admin/login");
 	}
-
+	
 	return <>{children}</>;
 }
