+++ src/components/auth/AuthProvider.tsx
@@ -1,11 +1,45 @@
 "use client";
 
-import { SessionProvider } from "next-auth/react";
+import { auth } from "@/lib/firebase";
+import { onAuthStateChanged } from "firebase/auth";
+import { createContext, useContext, useEffect, useState } from "react";
+
+// 認証コンテキストの型定義
+type AuthContextType = {
+  user: any | null;
+  loading: boolean;
+};
+
+// 認証コンテキストの作成
+const AuthContext = createContext<AuthContextType>({
+  user: null,
+  loading: true,
+});
+
+export function useAuth() {
+  return useContext(AuthContext);
+}
 
 export default function AuthProvider({
-	children,
+  children,
 }: {
-	children: React.ReactNode;
+  children: React.ReactNode;
 }) {
-	return <SessionProvider>{children}</SessionProvider>;
+  const [user, setUser] = useState<any | null>(null);
+  const [loading, setLoading] = useState(true);
+
+  useEffect(() => {
+    const unsubscribe = onAuthStateChanged(auth, (user) => {
+      setUser(user);
+      setLoading(false);
+    });
+
+    return () => unsubscribe();
+  }, []);
+
+  return (
+    <AuthContext.Provider value={{ user, loading }}>
+      {!loading && children}
+    </AuthContext.Provider>
+  );
 }
