+++ src/lib/firebase/store/profile.ts
@@ -1,40 +1,52 @@
 import { db, storage } from "../client";
-import { collection, getDocs, doc, getDoc } from "firebase/firestore";
+import { doc, getDoc } from "firebase/firestore";
 import { ref, uploadBytes, getDownloadURL } from "firebase/storage";
 import type { Profile } from "@/types/profile";
-import { auth } from "@/auth";
 
-export const getProfile = async (): Promise<Profile[]> => {
+
+export const getProfile = async (): Promise<Profile> => {
 	try {
-		const querySnapshot = await getDocs(collection(db, "profiles"));
+		const profilesRef = doc(db, "profiles", "main");
+		const docSnap = await getDoc(profilesRef);
+		
+		if (!docSnap.exists()) {
+			throw new Error("プロフィールデータが見つかりません");
+		}
 
-		return querySnapshot.docs.map((doc) => {
-			const data = doc.data();
-			return {
-				id: doc.id,
-				name: data.name,
-				nickname: data.nickname,
-				bio: data.bio,
-				imageUrl: data.imageUrl,
-				careers: data.careers,
-			} as Profile;
-		});
+		// 既存のデータから必要なフィールドのみを抽出
+		const data = docSnap.data();
+		return {
+			name: data.name,
+      		nickname: data.nickname,
+      		bio: data.bio,
+      		imageUrl: data.imageUrl,
+      		careers: data.careers || [],
+		} as Profile;
 	} catch (error) {
 		console.error("Error fetching profiles:", error);
 		throw error;
 	}
 };
 
+export async function getPublicProfile(): Promise<Profile | null> {
+	try {
+		const docRef = doc(db, "profiles", "main");
+		const docSnap = await getDoc(docRef);
+
+		if (docSnap.exists()) {
+			return docSnap.data() as Profile;
+		}
+		return null;
+	} catch (error) {
+		console.error("Error fetching profile:", error);
+		throw error;
+	}
+}
+
 // 画像をGCSにアップロードする関数
 export const uploadProfileImage = async (file: File): Promise<string> => {
 	if (!file) {
 		throw new Error("ファイルが選択されていません");
-	}
-
-	// Auth.jsのセッションを確認
-	const session = await auth();
-	if (!session) {
-		throw new Error("認証されていません。ログインしてください。");
 	}
 
 	try {
@@ -71,28 +83,3 @@ export const getProfileImageUrl = async (): Promise<string | null> => {
 		return null;
 	}
 };
-
-// 認証状態を確認する関数
-export const checkAuthState = async (): Promise<boolean> => {
-	const session = await auth();
-	console.log(
-		"Current auth state:",
-		session ? "authenticated" : "not authenticated"
-	);
-	return !!session;
-};
-
-export async function getPublicProfile(): Promise<Profile | null> {
-	try {
-		const docRef = doc(db, "profiles", "main");
-		const docSnap = await getDoc(docRef);
-
-		if (docSnap.exists()) {
-			return docSnap.data() as Profile;
-		}
-		return null;
-	} catch (error) {
-		console.error("Error fetching profile:", error);
-		throw error;
-	}
-}
