+++ src/types/profile.ts
@@ -2,13 +2,13 @@ export interface Profile {
 	name: string;
 	nickname: string;
 	bio: string;
-	imageUrl: string | null;
-	careers: Career[];
+	imageUrl?: string;
+	careers?: Career[];
 }
 
 export interface Career {
-	period: string;
 	description: string;
+	period: string;
 }
 
 export const PROFILE_CONSTRAINTS = {
